decimal128 and fix origin not in bounds/distorted grid issue

+++ src/mongo/db/exec/neighbors.h
@@ -48,7 +48,7 @@ namespace mongo {
 struct NeighborsStageParams {
     NeighborsStageParams() {}
     int k;
-    std::vector<double> queryPoint;
+    std::vector<Decimal128> queryPoint;
 
     // MatchExpression* filter; // TODO filter is too complicated. don't allow it for now
